compute_service/requirements.txt
Django==1.5
Pygments==1.6
ScientificPython==2.8
South==0.7.6
distribute==0.6.31
dj-database-url==0.2.1
djangorestframework==2.2.4
gunicorn==0.17.2
numpy==1.7.0
ply==3.4
psycopg2==2.4.6
wsgiref==0.1.2
